requirements.txt
pyautogui>=0.9.54
mss>=9.0.1
Pillow>=10.0.0
keyboard>=0.13.5
pystray>=0.19.5
python-dotenv>=1.0.0
requests>=2.31.0
pyperclip>=1.8.2
